Refactor regsync into a library
