+++ pkg/regsync/error.go
@@ -1,4 +1,4 @@
-package main
+package regsync
 
 import "errors"
